Don't process grid points with "small" X or Y coordinates.
Just like we scan for the highest X and Y coordinates, and don't
process any grid points with higher X and Y coordinates, we might
as well scan for the lowest X and Y coordinates, and don't process
any grid points whose X or Y coordinates are less than that.

Saves about 25% of processing time.