A real time solution for Day 7, part 2.
Here we actually let elves work, for the given duration. Each
working elf will be a child process, and we only spawn a new
child if we have an elf available, and a component to be worked
on. The total running time will be the answer to the question.

Caveats: Two things may spoil it: the overhead of assigning a task,
         which in the challenge is given as 0 time. And we're using
         sleep() which isn't guaranteed to be exact.

The process won't be busy waiting; waitpid() behaves nicely.