Replaced brute-force solution by a much faster one.
We now create partial sums; that is, for each point (x,y) store
the sum of all points (x',y') with x' <= x and y' <= y. This can
be efficiently calculated:

  ps (x, y) = value (x, y) + ps (x - 1, y) + ps (x, y - 1) - ps (x - 1, y - 1)

The sum of the values of a rectangle bounded by (x1, y1), (x2, y2)
(with x1 <= x2 and y1 <= y2) is then

  ps (x2, y2) - ps (x2, y1 - 1) - ps (x1 - 1, y2) + ps (x1 - 1, y1 - 1)

This reduced the running time from 2m24s to 4s.